requirements.txt
asgiref==3.6.0
attrs==22.2.0
boto3==1.26.62
botocore==1.29.62
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cloudinary==1.31.0
cryptography==39.0.0
defusedxml==0.7.1
dj-database-url==1.2.0
dj-email-url==1.0.6
Django==4.1.4
django-allauth==0.52.0
django-cache-url==3.4.4
django-cors-headers==3.13.0
django-extensions==3.2.1
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
easy-thumbnails==2.8.4
environs==9.5.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.17.3
MarkupSafe==2.1.1
marshmallow==3.19.0
oauthlib==3.2.2
packaging==22.0
Pillow==9.3.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.7
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
Werkzeug==2.2.2
whitenoise==6.2.0
